+++ crates/biome_configuration/src/analyzer/linter/rules.rs
@@ -2981,12 +2981,12 @@ pub struct Nursery {
         Option<RuleConfiguration<biome_css_analyze::options::NoUnknownProperty>>,
     #[doc = "Disallow unknown pseudo-class selectors."]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub no_unknown_pseudo_class_selector:
-        Option<RuleConfiguration<biome_css_analyze::options::NoUnknownPseudoClassSelector>>,
+    pub no_unknown_pseudo_class:
+        Option<RuleConfiguration<biome_css_analyze::options::NoUnknownPseudoClass>>,
     #[doc = "Disallow unknown pseudo-element selectors."]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub no_unknown_selector_pseudo_element:
-        Option<RuleConfiguration<biome_css_analyze::options::NoUnknownSelectorPseudoElement>>,
+    pub no_unknown_pseudo_element:
+        Option<RuleConfiguration<biome_css_analyze::options::NoUnknownPseudoElement>>,
     #[doc = "Disallow unknown CSS units."]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub no_unknown_unit: Option<RuleConfiguration<biome_css_analyze::options::NoUnknownUnit>>,
@@ -3152,8 +3152,8 @@ impl Nursery {
         "noUnknownFunction",
         "noUnknownMediaFeatureName",
         "noUnknownProperty",
-        "noUnknownPseudoClassSelector",
-        "noUnknownSelectorPseudoElement",
+        "noUnknownPseudoClass",
+        "noUnknownPseudoElement",
         "noUnknownUnit",
         "noUnmatchableAnbSelector",
         "noUnusedFunctionParameters",
@@ -3201,8 +3201,8 @@ impl Nursery {
         "noShorthandPropertyOverrides",
         "noUnknownFunction",
         "noUnknownProperty",
-        "noUnknownPseudoClassSelector",
-        "noUnknownSelectorPseudoElement",
+        "noUnknownPseudoClass",
+        "noUnknownPseudoElement",
         "noUnknownUnit",
         "noUnmatchableAnbSelector",
         "noUselessEscapeInRegex",
@@ -3466,12 +3466,12 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[28]));
             }
         }
-        if let Some(rule) = self.no_unknown_pseudo_class_selector.as_ref() {
+        if let Some(rule) = self.no_unknown_pseudo_class.as_ref() {
             if rule.is_enabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[29]));
             }
         }
-        if let Some(rule) = self.no_unknown_selector_pseudo_element.as_ref() {
+        if let Some(rule) = self.no_unknown_pseudo_element.as_ref() {
             if rule.is_enabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[30]));
             }
@@ -3780,12 +3780,12 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[28]));
             }
         }
-        if let Some(rule) = self.no_unknown_pseudo_class_selector.as_ref() {
+        if let Some(rule) = self.no_unknown_pseudo_class.as_ref() {
             if rule.is_disabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[29]));
             }
         }
-        if let Some(rule) = self.no_unknown_selector_pseudo_element.as_ref() {
+        if let Some(rule) = self.no_unknown_pseudo_element.as_ref() {
             if rule.is_disabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[30]));
             }
@@ -4097,12 +4097,12 @@ impl Nursery {
                 .no_unknown_property
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
-            "noUnknownPseudoClassSelector" => self
-                .no_unknown_pseudo_class_selector
+            "noUnknownPseudoClass" => self
+                .no_unknown_pseudo_class
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
-            "noUnknownSelectorPseudoElement" => self
-                .no_unknown_selector_pseudo_element
+            "noUnknownPseudoElement" => self
+                .no_unknown_pseudo_element
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
             "noUnknownUnit" => self